+++ apps/nextjs/src/app/(routes)/score-card/[id]/use-columns.tsx
